<Feature, UI> New Button Style, Button Hold
Allow subscription to OnDown, OnHold, and OnUp events for all axis on
buttons.
Menu Buttons have bookends, game buttons have underscores.
Inventory items can be moved in batches of 5.
Moved Home into WorldState.
Switched to using integer values for resource quantities.

+++ Night Unity Files/Assets/Scripts/Facilitating/Persistence/GameData.cs
@@ -18,7 +18,7 @@ namespace Facilitating.Persistence
         public static List<Character> Party;
         
         //Camp Data
-        public static float StoredWater = 0f, StoredFood = 0f, StoredFuel = 0f;
+        public static int StoredWater = 0, StoredFood = 0, StoredFuel = 0;
 
         public static void SetDifficultyFromString(string difficultyString)
         {
