+++ Night Unity Files/Assets/Scripts/Game/World/Region/RegionTemplate.cs
@@ -3,6 +3,6 @@
     public class RegionTemplate
     {
         public string InternalName, DisplayName, Type, Encounters, Items;
-        public float WaterAvailable, FoodAvailable, FuelAvailable, ScrapAvailable, AmmoAvailable;
+        public int WaterAvailable, FoodAvailable, FuelAvailable, ScrapAvailable, AmmoAvailable;
     }
 }
