+++ Night Unity Files/Assets/Scripts/SamsHelper/BaseGameFunctionality/InventorySystem/BasicInventoryContents.cs
@@ -21,9 +21,9 @@
             return _weight;
         }
 
-        public virtual float Quantity()
+        public virtual int Quantity()
         {
-            return 1f;
+            return 1;
         }
     }
 }
